Add first smoke and API tests
Create `test/lab/*.js` which provides the laboratory environment for
executing tests:
 - node-tap as the test runner
 - chai + dirty-chai for assertions
 - Bluebird for promises

Add a smoke test to verify that the native add-on can be loaded
and the `start` method can be successfully executed.

Add tests for the public function `start(port, cb)`, verify that
incorrect arguments are rejected with a helpful error message.