Add periodic call of `ProcessDebugMessages()`
When V8 is not running (e.g. because libuv is waiting for an event),
it cannot process debugger commands. The solution is to periodically
call `v8::Debug::ProcessDebugMessages()`.

In strong-debugger, we know exactly when a new command has arrived,
thus we can immediatelly schedule ProcessDebugMessages() on the main
event loop.